Deep link support for quick access to habits and daily logs
Dedicated habit statistics screen with streak tracking
Backup and restore data functionality
Video capture and playback support in daily logs
Habit sorting options (today's habits vs. all habits)
System font preference option
Separate screens for Habit and Daily Log.
Media support: video playback, multiple images per daily log, full-screen image viewer, and media sharing/downloading.
Enhanced daily log media carousel with multiple images and videos
Improved notification actions and habit completion tracking
Hindi language support added
Updated app shortcuts for quick actions
Better UI components and visual consistency

+++ app/src/main/java/com/charan/habitdiary/presentation/add_habit/AddHabitState.kt
@@ -18,7 +18,7 @@ data class AddHabitState(
     val isReminderEnabled : Boolean = false,
     val showPermissionRationale : Boolean = false,
     val showDeleteDialog : Boolean = false,
-    val habitId : Int? = null,
+    val habitId : Long? = null,
     val isEdit : Boolean = false,
     val is24HourFormat : Boolean = false
 )
